pyproject.toml
[tool.poetry]
package-mode = false
name = "isaco"
version = "0.1.0"
description = ""
authors = ["Hadi Mohseni <hadimohseni1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyodbc = "^5.2.0"
python-dotenv = "^1.0.1"
awesometkinter = "^2021.11.8"
tkintertable = "^1.3.3"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
joblib = "^1.4.2"
customtkinter = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
